+++ CHANGELOG.md
@@ -6,7 +6,8 @@
 - Handle logout when token is invalid.
 - Possibility to run TMC Actions for exercises when right clicking exercise folders or files in the tree view.
     * This way .ipynb notebooks can be tested and submitted more easily.
-- Clear new exercises possibility from notification or course card.
+- Buttons to clear new exercises buffer from notification or course card.
+- Inform user if TMC Extension is obsolete.
 
 #### Changed
 - Technical improvements to Login page.